src/test/java/ru/begletsov/job4j_tasks/_116_arrays/ValidatorUniqueElements_33Test.java
package ru.begletsov.job4j_tasks._116_arrays;

import org.junit.Test;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;

public class ValidatorUniqueElements_33Test {
    @Test
    public void checkArrayFalse() {
        int[] data = {1, 2, 1, 3, 1, 4, 1};
        int value = 1;
        boolean rsl = ValidatorUniqueElements_33.checkArray(data, value);
        assertThat(rsl, is(false));
    }

    @Test
    public void checkArrayTrue() {
        int[] data = {1, 2, 3, 1, 4, 1, 5, 6, 7, 1, 8, 9};
        int value = 1;
        boolean rsl = ValidatorUniqueElements_33.checkArray(data, value);
        assertThat(rsl, is(true));
    }

}